[CDF-22685] 😼 Builder Concept Part 3 (#1033)
* build; ignore folder

* refactor; added builders package

* refactor: setup shell for builder

* refactor: moved all code

* refactor; moved the rest

* refactor: switch over

* tests: updated tets

* refactor; cleanup double use

* tests: changed tests

* refactor; fix removed needle folder

* refactor: moved out function builder

* refactor: moved out file builder

* refactor: fix

* refactor: moved out into seperate files

* refactor: moved setup spec

* refactor; moved in destination

* refactor: move around data

* refactor: moved out deploy results

* refactor; moved data classes out of loaders

* refactor; split Raw DataBase and Table

* refactor: replacing RawDataBaseTable

* fix: validation to assume either RawDabase or RawTable
