+++ cognite_toolkit/_cdf_tk/commands/clean.py
@@ -19,6 +19,9 @@ from cognite_toolkit._cdf_tk.constants import (
 )
 from cognite_toolkit._cdf_tk.data_classes import (
     BuildEnvironment,
+    DeployResults,
+    ResourceContainerDeployResult,
+    ResourceDeployResult,
 )
 from cognite_toolkit._cdf_tk.exceptions import (
     ToolkitCleanResourceError,
@@ -28,12 +31,10 @@ from cognite_toolkit._cdf_tk.loaders import (
     LOADER_BY_FOLDER_NAME,
     DataLoader,
     DataSetsLoader,
-    DeployResults,
     ResourceContainerLoader,
     ResourceLoader,
 )
 from cognite_toolkit._cdf_tk.loaders._base_loaders import T_ID, Loader, T_WritableCogniteResourceList
-from cognite_toolkit._cdf_tk.loaders.data_classes import ResourceContainerDeployResult, ResourceDeployResult
 from cognite_toolkit._cdf_tk.tk_warnings import (
     LowSeverityWarning,
     MediumSeverityWarning,
