+++ cognite_toolkit/_cdf_tk/loaders/_data_loaders.py
@@ -8,11 +8,11 @@ from typing import TYPE_CHECKING, cast, final
 import pandas as pd
 from cognite.client.data_classes import FileMetadataWrite, FileMetadataWriteList
 
+from cognite_toolkit._cdf_tk.client.data_classes.raw import RawTable
 from cognite_toolkit._cdf_tk.utils import CDFToolConfig, read_yaml_content, safe_read
 
 from ._base_loaders import DataLoader
 from ._resource_loaders import FileMetadataLoader, RawTableLoader, TimeSeriesLoader
-from .data_classes import RawDatabaseTable
 
 if TYPE_CHECKING:
     from cognite_toolkit._cdf_tk.data_classes import BuildEnvironment, BuiltResource
@@ -142,7 +142,7 @@ class RawFileLoader(DataLoader):
         for resource in state.built_resources[self.folder_name]:
             if resource.kind != RawTableLoader.kind:
                 continue
-            table = cast(RawDatabaseTable, resource.identifier)
+            table = cast(RawTable, resource.identifier)
             datafile = next(
                 (
                     resource.location.path.with_suffix(f".{file_type}")
