+++ tests/test_unit/test_cdf_tk/test_loaders/test_data_model_loader.py
@@ -6,11 +6,11 @@ from unittest.mock import MagicMock
 import pytest
 from cognite.client.data_classes import data_modeling as dm
 
+from cognite_toolkit._cdf_tk.client.data_classes.graphql_data_models import GraphQLDataModel
 from cognite_toolkit._cdf_tk.commands import DeployCommand
 from cognite_toolkit._cdf_tk.exceptions import ToolkitCycleError
 from cognite_toolkit._cdf_tk.loaders import DataModelLoader
 from cognite_toolkit._cdf_tk.loaders._resource_loaders import GraphQLLoader
-from cognite_toolkit._cdf_tk.loaders.data_classes import GraphQLDataModel
 from cognite_toolkit._cdf_tk.utils import CDFToolConfig
 from tests.test_unit.approval_client import ApprovalToolkitClient
